+++ app/controllers/projects_controller.rb
@@ -1,7 +1,7 @@
 class ProjectsController < ApplicationController
   before_action :authenticate_user!
   before_action :check_system_admin #, only: [ :new, :create, :edit, :update, :destroy ]
-  before_action :set_project, only: [:show, :logo, :edit, :update, :destroy]
+  before_action :set_project, only: [:show, :agency_logo, :client_logo, :edit, :update, :destroy]
 
   layout 'application'
 
@@ -16,8 +16,12 @@ class ProjectsController < ApplicationController
   def show
   end
 
-  def logo
-    send_file File.join( CarrierWave::Uploader::Base.root, @project.logo.url )
+  def agency_logo
+    send_file File.join( CarrierWave::Uploader::Base.root, @project.agency_logo.url )
+  end
+
+  def client_logo
+    send_file File.join( CarrierWave::Uploader::Base.root, @project.client_logo.url )
   end
 
   # GET /projects/new
@@ -82,6 +86,10 @@ class ProjectsController < ApplicationController
     def project_params
       params[:project] ||= { blank: '1' }
       params[:project][:password_plain] = params[:project][:password] unless params[:project][:password].blank?
-      params.require(:project).permit(:name, :number, :slug, :logo, :logo_cache, :client_name, :username, :password, :password_plain)
+      params.require(:project).permit(
+        :name, :number, :slug,
+        :agency_logo, :agency_logo_cache,
+        :client_logo, :client_logo_cache,
+        :client_name, :username, :password, :password_plain)
     end
 end