+++ db/migrate/20150623230857_change_logo_to_agency_logo.rb
@@ -1,0 +1,5 @@
+class ChangeLogoToAgencyLogo < ActiveRecord::Migration
+  def change
+    rename_column :projects, :logo, :agency_logo
+  end
+end