+++ db/migrate/20150623231421_add_client_logo_to_projects.rb
@@ -1,0 +1,5 @@
+class AddClientLogoToProjects < ActiveRecord::Migration
+  def change
+    add_column :projects, :client_logo, :string
+  end
+end