+++ test/controllers/projects_controller_test.rb
@@ -19,7 +19,7 @@ class ProjectsControllerTest < ActionController::TestCase
 
   test "should create project" do
     assert_difference('Project.count') do
-      post :create, project: { name: 'New Project', slug: 'new-project', logo: @project.logo, password: @project.password_plain }
+      post :create, project: { name: 'New Project', slug: 'new-project', agency_logo: @project.agency_logo, password: @project.password_plain }
     end
 
     assert_redirected_to project_path(assigns(:project))
@@ -36,7 +36,7 @@ class ProjectsControllerTest < ActionController::TestCase
   end
 
   test "should update project" do
-    patch :update, id: @project, project: { name: 'Project Update', slug: 'project-update', logo: @project.logo, password: @project.password_plain }
+    patch :update, id: @project, project: { name: 'Project Update', slug: 'project-update', agency_logo: @project.agency_logo, password: @project.password_plain }
     assert_redirected_to project_path(assigns(:project))
   end